requirements.txt
paho-mqtt==1.6.1
pymongo==4.6.2
python-dotenv==1.0.1 